chore(vat-data): move `vow.js` to `@agoric/vow/vat.js`
