+++ packages/vat-data/package.json
@@ -30,7 +30,6 @@
   },
   "devDependencies": {
     "@endo/init": "^1.1.1",
-    "@endo/far": "^1.1.1",
     "@endo/ses-ava": "^1.2.1",
     "ava": "^5.3.0",
     "tsd": "^0.30.7"
