+++ packages/vow/package.json
@@ -21,6 +21,7 @@
   },
   "dependencies": {
     "@agoric/base-zone": "^0.1.0",
+    "@agoric/internal": "^0.3.2",
     "@endo/env-options": "^1.1.3",
     "@endo/eventual-send": "^1.2.1",
     "@endo/pass-style": "^1.3.1",
@@ -28,7 +29,6 @@
     "@endo/promise-kit": "^1.1.1"
   },
   "devDependencies": {
-    "@agoric/internal": "^0.3.2",
     "@endo/far": "^1.1.1",
     "@endo/init": "^1.1.1",
     "ava": "^5.3.0"
